add: middleware

+++ laragigs/routes/web.php
@@ -18,20 +18,20 @@ use Illuminate\Support\Facades\Route;
 
 // Listing
 Route::get('/', [ListingController::class, 'index']);
-Route::post('/listings', [ListingController::class, 'store']);
-Route::get('/listings/create', [ListingController::class, 'create']);
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 
 // User
-Route::post('/users', [UserController::class, 'store']);
-Route::post('/login', [UserController::class, 'login']);
-Route::get('/login', [UserController::class, 'getLogin']);
-Route::post('/logout', [UserController::class, 'logout']);
-Route::get('/register', [UserController::class, 'create']);
+Route::post('/users', [UserController::class, 'store'])->middleware('guest');
+Route::post('/login', [UserController::class, 'login'])->middleware('guest');
+Route::get('/login', [UserController::class, 'getLogin'])->name('login')->middleware('guest');
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
